posix_acl: Clear SGID bit when setting file permissions
commit 073931017b49d9458aa351605b43a7e34598caef upstream.

When file permissions are modified via chmod(2) and the user is not in
the owning group or capable of CAP_FSETID, the setgid bit is cleared in
inode_change_ok().  Setting a POSIX ACL via setxattr(2) sets the file
permissions as well as the new ACL, but doesn't clear the setgid bit in
a similar way; this allows to bypass the check in chmod(2).  Fix that.

References: CVE-2016-7097
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

# Conflicts:
#	fs/9p/acl.c
#	fs/btrfs/acl.c
#	fs/ext2/acl.c
#	fs/ext4/acl.c
#	fs/f2fs/acl.c
#	fs/gfs2/acl.c
#	fs/hfsplus/posix_acl.c
#	fs/ocfs2/acl.c
#	fs/reiserfs/xattr_acl.c
#	fs/xfs/xfs_acl.c
#	include/linux/posix_acl.h